requirements.txt
aiohttp
python-dotenv
python-telegram-bot
websockets
websocket-client